Add uncached tree hashing
