+++ eth2/utils/tree_hash/src/cached_tree_hash/resize.rs
